+++ ingester2/src/wal/reference_tracker/mod.rs
@@ -1,0 +1,8 @@
+mod actor;
+mod handle;
+mod wal_deleter;
+
+pub(crate) use actor::*;
+#[allow(unused_imports)] // Used by docs - will be used by code soon.
+pub(crate) use handle::*;
+pub(crate) use wal_deleter::*;
